requirements.txt
cachetools==3.1.0
certifi==2019.3.9
chardet==3.0.4
Click==7.0
Flask==1.0.2
Flask-Cors==3.0.7
google-api-core==1.11.0
google-auth==1.6.3
google-cloud-core==1.0.0
google-cloud-storage==1.15.1
google-resumable-media==0.3.2
googleapis-common-protos==1.5.10
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
protobuf==3.7.1
pyasn1==0.4.5
pyasn1-modules==0.2.5
pytz==2019.1
requests==2.22.0
rethinkdb==2.4.2.post1
rsa==4.0
six==1.12.0
sox==1.3.7
urllib3==1.25.2
Werkzeug==0.15.4
youtube-dl==2019.5.11
